backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2
twilio==8.10.0
elevenlabs==0.2.26
google-cloud-translate==3.12.1
